Drastic model overhaul.
The model now stores instances of SeasidePerson, a QObject type on which
properties are set. This allows QML UIs to easily gain change notifications on
properties, as well as easily enabling sub-UIs which don't have access to the
model itself to monitor for changes on a given contact instance.

This is probably also a lot more performant, from looking at some of the
previous code, as it doesn't constantly rebuild QContacts...